modify functions location to be reuseable
